Updated query pagination logic in client.py to utilize yield statements while iterating over query results.
When performing a query the number of expected pages is now calculated from the total hits, and used to determine when pagination has been exhausted.
Additionally, a check to ensure a Products object query clause is not modified while mid-pagination has been added.
A new method called reset() can be used to clear pagination status to return the Products object to a default state for a new query.