Added the products_with_target() method to QueryBuilder
This method adds a LID target filter to the current query string based on a user-provided keyword.
The _add_clause() method has also been updated to accept a logical_join argument that instructs how a new query clause should be logically joined with the rest of the query.
This feature will be reworked in a future branch to properly handle assembly of sub-clauses with the appopriate logical operator.